requirements.txt
aiogram<3.0
openai
python-dotenv
async-openai
